Minor fixes to ensure that bool values are cast pr
operly during scheduled sampling

+++ espnet/nets/pytorch_backend/transducer/joint_network.py
@@ -128,7 +128,7 @@ class JointNetwork(torch.nn.Module):
                 for b in range(am_outs.shape[0]):
                     for t in range(T):
                         la_tokens[b,t] = torch.cat([am_outs[b,t+1:][am_outs[b,t+1:]!=0][:self.la_window],torch.zeros(self.la_window,device=enc_out.device,dtype=am_outs.dtype)])[:self.la_window]
-                la_tokens =  la_tokens.unsqueeze(-2).expand(-1,-1,U,-1)   # Shape here is B x T x 1 x embed*num_tokens
+                la_tokens =  la_tokens.unsqueeze(-2).expand(-1,-1,U,-1)   # Shape here is B x T x U x embed*num_tokens
                 if self.training:  # Perform scheduled sampling only during training
                     target = torch.cat([target,torch.zeros([B,1],device=am_outs.device,dtype=target.dtype)],dim=-1)
                     sched_samp = torch.zeros([B,U,self.la_window],dtype=am_outs.dtype,device=am_outs.device)
@@ -137,7 +137,7 @@ class JointNetwork(torch.nn.Module):
                             sched_samp[b,u] = torch.cat([target[b,u:][:self.la_window],torch.zeros(self.la_window,device=enc_out.device,dtype=am_outs.dtype)])[:self.la_window]
                     sched_samp = sched_samp.unsqueeze(1).expand(-1,T,-1,-1)
                     sched_samp_rand = torch.rand(sched_samp.shape,device=la_tokens.device)
-                    la_tokens = la_tokens * (sched_samp_rand > self.la_greedy_scheduled_sampling_probability).to(int) + sched_samp * (sched_samp_rand <= self.la_greedy_scheduled_sampling_probability)
+                    la_tokens = la_tokens * (sched_samp_rand > self.la_greedy_scheduled_sampling_probability).to(int) + sched_samp * (sched_samp_rand <= self.la_greedy_scheduled_sampling_probability).to(int)
                 la_tokens = self.embed_la(la_tokens).reshape(B,T,U,-1)
                 dec_out = dec_out.expand(-1,T,-1,-1)
                 dec_out = torch.cat([dec_out,la_tokens],dim=-1)
